// sdk/durabletask/arm-durabletask/src/models/models.ts
export interface Resource {
  /** Fully qualified resource ID for the resource. Ex - /subscriptions/{subscriptionId}/resourceGroups/{resourceGroupName}/providers/{resourceProviderNamespace}/{resourceType}/{resourceName} */
  readonly id?: string;
  /** The name of the resource */
  readonly name?: string;
  /** The type of the resource. E.g. "Microsoft.Compute/virtualMachines" or "Microsoft.Storage/storageAccounts" */
  readonly type?: string;
  /** Azure Resource Manager metadata containing createdBy and modifiedBy information. */
  readonly systemData?: SystemData;
}
// ...
export interface ProxyResource extends Resource {}

export function proxyResourceSerializer(item: ProxyResource) {
  return item as any;
}

/** A Task Hub resource belonging to the scheduler */
export interface TaskHub extends ProxyResource {
  /** The resource-specific properties for this resource. */
  properties?: TaskHubProperties;
}

export function taskHubSerializer(item: TaskHub): Record<string, unknown> {
  return {
    properties: !item.properties
      ? item.properties
      : taskHubPropertiesSerializer(item.properties),
  };
}

/** The properties of Task Hub */
export interface TaskHubProperties {
  /** The status of the last operation */
  readonly provisioningState?: ProvisioningState;
  /** URL of the durable task scheduler dashboard */
  readonly dashboardUrl?: string;
}
// ...
export interface _TaskHubListResult {
  /** The TaskHub items on this page */
  value: TaskHub[];
  /** The link to the next page of items */
  nextLink?: string;
}
// ...
export type ProvisioningState =
  | string
  | ResourceProvisioningState
  | "Provisioning"
  | "Updating"
  | "Deleting"
  | "Accepted";